Rename einsum.py to einsumfuncs.py + noqa import
